+++ packages/react/src/modules/guide/components/types.ts
@@ -1,5 +1,19 @@
 import { KnockGuide, KnockGuideStep } from "@knocklabs/client";
 
+export type ImageContent = {
+  url: string;
+  alt: string;
+  action: string;
+};
+
+export type TargetImage = ImageContent;
+
+export type TargetImageWithGuide = {
+  image: ImageContent;
+  step: KnockGuideStep;
+  guide: KnockGuide;
+};
+
 export type ButtonContent = {
   text: string;
   action: string;
@@ -9,7 +23,7 @@ export type TargetButton = ButtonContent & {
   name: string;
 };
 
-export type TargetButtonWithGuideContext = {
+export type TargetButtonWithGuide = {
   button: TargetButton;
   step: KnockGuideStep;
   guide: KnockGuide;